test: Serve endpoint provides tc value for cache misses
During the analysis of export targets, if there is a local target
cache miss, we should also interrogate the serve endpoint
(if given) for the target cache value before proceeding with the
analysis.

For this test we have extended the "with_serve" runner to expose
also the remote-execution build root to the test env.