just-mr: don't make the command line pretend a fixed name
Our fetch and launch tool is parametric in the tool to be launched.
Reflect this in the documentation and do not pretend it to be the
name "just" hard coded. While there, also fix the hard-coded name
"git" in the documentation of the default value.

(cherry picked from commit e15f11ce0a7813602d9627e95ae9fb125afa891e)