fpath_git_map: Sync pragma 'to_git' root trees with serve endpoint
Marking a file-type repository as 'to_git' results in a Git-tree
type root, which are of course content fixed and can be (and
usually are) used by export targets. Therefore, it is beneficial
for a serve endpoint, if one is provided, to be aware of such a
root and be able to build against it if needed. If the root is
marked as absent, this condition becomes mandatory.

Generating an absent Git-tree root without being provided a serve
endpoint is still allowed, but results in a warning.