Design interaction between target-level caching and garbage collection
The implicit extensional projection of target-level caching and
garbage collection interact in a subtile way. Add a design document
describing how we keep the invariant required by target-level
caching in the presence of garbage collection. While, techincally,
this just describes how to fix a bug, a careful design is still
needed, as some fundamental changes are made.