just-mr 'git tree' repository: Absent roots sync with serve endpoint
To take advantage of absent roots, we need to ensure that a given
serve endpoint can build against the tree of this generated root.

To this end, for a 'git tree' repository we only set the root as
absent only if the given serve endpoint has this root, or the tree
is known locally and can be provided via the remote CAS. While
generating an absent root the fetch command will never be called.

Generating an absent root without being provided a serve endpoint
is still allowed, but results in a warning.