just-mr 'git' repository: Absent roots sync with serve endpoint
To take advantage of absent roots, we need to ensure that a given
serve endpoint can build against the tree of this generated root.

To this end, for a 'git' repository we only set the root as absent
if the serve endpoint knows already the root, it can set it up
itself, or we can create the root locally without a network fetch
and then upload it to the serve endpoint via the remote CAS.
A network fetch never gets performed for an absent root.

If a serve endpoint is not provided, an absent root can still be
generated, but only if no network fetches are required. In this
case a warning is emitted.