just serve: Improve progress reporting
During analysis it is useful to track and report the progress for
all export targets. This is not exclusively linked to a serve
endpoint being present, despite most of the time being expected to
be spent in export targets being served from the remote endpoint.

This commit refactors the current implementation to give proper
feedback to the user on the progress of the analysis phase.