bug fix in expresion, Union: propagate the disjointness property
To avoid too many intermediate results, we compute the union of
a list in a divide and conquer fashion. Of course, for a disjoint
union, the recursive calls on the lists of half the length have to
be disjoint as well, i.e., the template parameter kDisjoint has to
be passed on. Fix this.

(cherry picked from commit 469a7190bee490344d147099c3511fc45e2aadf0)