grpc: hide dependency on google_rpc_status
grpc is used in the toolchain defaults for proto servive libraries.
Still, it is typically built on its own, with its own toolchain,
flags, etc. Now, grpc, however has a public dependency on a the
rpc-status proto library, that the user may well use on their own,
however building in their own way which can yield conflicts. To solve
this, we hide the dependency on that proto library, as infrastructure
libs should not make assumptions on user-servicable libraries.
- First, we note that the dependency can be made a private one,
  which already solves the conflict on header files (which will
  essentially be the same, but might be defined in a different way).
- Next, we note that the library at linking basically only acts
  as a default implementation; if the user provides their own
  version of the rpc-status library, we should prefer that anyway.
  As infrastructure is linked last, we have that default character
  anyway; the only thing to do is to rename the library that no
  staging conflict occurs.