just-mr: Ensure resolved trees are kept alive in Git cache
The association map file for a resolved tree was supposed to
guarantee that the respective tree is kept alive in a Git
repository as part of a tagged commit.

This commit fixes this issue by tagging the tree (found in the Git
cache after resolution) before writing its association file.