describe: also describe "configure" targets in more detail
Most built-in rules are just described by their name, assuming they
are known to the user anyway. One exception, however, are "export"
targets, as those serve as interfaces to (logical) repositories and
hence is supposed to have a documentation of the target itself. Over
time, however, "configure" targets have evolved to become internal
interfaces (maybe even for an external target), e.g., to provide
default values and hence make better use of the actual export target.
Therefore it is desirable to provide a bit more information when
asked to describe a "configure" target.
- As "configure" targets have a fixed set of keys, adding a "doc"
  field is a conservative extension; this can be useful to give an
  overview what the target is about.
- An important information of a "configure" target is the target
  that is configured. While this, in general, is an expression, in
  the typical cases, the description is very short (a literal target
  name, or a variable). So we can afford to show the definition.