TargetCache: Use StorageConfig instance for sharding
Instead of computing the shard based on the RemoteExecutionConfig
singleton, use the already computed hash stored in the passed
StorageConfig instance, which now needs to be set up separately
if bootstrapping in order to avoid unwanted includes.

Storing the backend description to CAS and corresponding audit
checks now take place in main.